Continue grpc things (#35)
* update openapi endpoint docs

* add account model proto

* fill in service getaccount types

* generate

* improve(?) pagination for generalized values (grpc+http)

* first iter proto idl

* transform account->pbaccount, basis as enum, transformers

* scaffold out grpc service method implementation

* export enum slug, implement grpc create methods

because the current implementation of go-playground/validtor cannot inspect unexported fields we're now exporting the enum slugs such that validation can be perform upon them. This is super unfortunate. Dean Kearn acknowledged that it may be possible to do so a couple years ago. There seems to be a small bit of action on this issue recently. Because the dependent code is just internal, I feel comfortable exporting it, however it opens up contributions to make mistakes.

* wire up grpc server; fix liablity typo error